+++ controllers/resource/resource_generator_test.go
@@ -2,6 +2,7 @@ package resource
 
 import (
 	_ "embed"
+	corev1 "github.com/cloudogu/k8s-dogu-operator/api/v1"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -301,5 +302,69 @@ func Test_createLivenessProbe(t *testing.T) {
 
 		// then
 		require.Nil(t, actual)
+	})
+}
+
+func Test_getChownInitContainer(t *testing.T) {
+	t.Run("success with whitespace in volume path", func(t *testing.T) {
+		// given
+		dogu := &core.Dogu{Volumes: []core.Volume{{Name: "whitespace", Path: "/etc/ldap config/test", Owner: "100", Group: "100"}}}
+		doguResource := &corev1.Dogu{ObjectMeta: metav1.ObjectMeta{Name: "ldap"}}
+		expectedCommand := []string{"sh", "-c", "mkdir -p \"/etc/ldap config/test\" && chown -R 100:100 \"/etc/ldap config/test\""}
+
+		// when
+		container, err := getChownInitContainer(dogu, doguResource)
+
+		// then
+		require.NoError(t, err)
+		require.Equal(t, expectedCommand, container.Command)
+	})
+
+	t.Run("should return nil if volumes are only of type dogu-operator", func(t *testing.T) {
+		// given
+		dogu := &core.Dogu{Volumes: []core.Volume{{Clients: []core.VolumeClient{{Name: "k8s-dogu-operator"}}}}}
+
+		// when
+		container, err := getChownInitContainer(dogu, nil)
+
+		// then
+		require.NoError(t, err)
+		require.Nil(t, container)
+	})
+
+	t.Run("should return error if owner cannot be parsed", func(t *testing.T) {
+		// given
+		dogu := &core.Dogu{Volumes: []core.Volume{{Name: "test", Owner: "3sdf"}}}
+
+		// when
+		_, err := getChownInitContainer(dogu, nil)
+
+		// then
+		require.Error(t, err)
+		assert.ErrorContains(t, err, "failed to parse owner id 3sdf from volume test")
+	})
+
+	t.Run("should return error if group cannot be parsed", func(t *testing.T) {
+		// given
+		dogu := &core.Dogu{Volumes: []core.Volume{{Name: "test", Owner: "1", Group: "3sdf"}}}
+
+		// when
+		_, err := getChownInitContainer(dogu, nil)
+
+		// then
+		require.Error(t, err)
+		assert.ErrorContains(t, err, "failed to parse group id 3sdf from volume test")
+	})
+
+	t.Run("should return error if ids are not greater than 0", func(t *testing.T) {
+		// given
+		dogu := &core.Dogu{Volumes: []core.Volume{{Name: "test", Owner: "0", Group: "-1"}}}
+
+		// when
+		_, err := getChownInitContainer(dogu, nil)
+
+		// then
+		require.Error(t, err)
+		assert.ErrorContains(t, err, "owner 0 or group -1 are not greater than 0")
 	})
 }